Relocate datasources logic
