+++ src/op_analytics/dagster/assets/github.py
@@ -3,7 +3,7 @@ from dagster import OpExecutionContext, asset
 
 @asset
 def traffic(context: OpExecutionContext) -> None:
-    from op_analytics.cli.subcommands.pulls.github import execute
+    from op_analytics.datasources.github import execute
 
     result = execute.execute_pull_traffic()
     context.log.info(result)
@@ -11,7 +11,7 @@ def traffic(context: OpExecutionContext) -> None:
 
 @asset
 def activity(context: OpExecutionContext) -> None:
-    from op_analytics.cli.subcommands.pulls.github import execute
+    from op_analytics.datasources.github import execute
 
     result = execute.execute_pull_activity()
     context.log.info(result)
@@ -19,7 +19,7 @@ def activity(context: OpExecutionContext) -> None:
 
 @asset(deps=[traffic, activity])
 def write_to_clickhouse(context: OpExecutionContext) -> None:
-    from op_analytics.cli.subcommands.pulls.github import execute
+    from op_analytics.datasources.github import execute
 
     result = execute.insert_to_clickhouse()
     context.log.info(result)
