+++ src/op_analytics/datasources/defillama/stablecoins.py
