+++ src/op_analytics/datasources/defillama/stablecoins_bigquery.py
