+++ src/op_analytics/datasources/github/execute.py
@@ -1,7 +1,7 @@
 from op_analytics.coreutils.logger import structlog
 
 from .activity.allrepos import GithubActivityData
-from .dataaccess import GithubDataset
+from .dataaccess import Github
 from .traffic.allrepos import GithubTrafficData
 from .traffic.bigquery import write_traffic_to_bq
 
@@ -14,12 +14,12 @@ def execute_pull_traffic():
     summary = {}
     summary["bigquery"] = write_traffic_to_bq(data)
 
-    GithubDataset.TRAFFIC_METRICS.write(
+    Github.TRAFFIC_METRICS.write(
         dataframe=data.all_metrics_df_truncated,
         sort_by=["dt", "metric"],
     )
 
-    GithubDataset.REFERRER_METRICS.write(
+    Github.REFERRER_METRICS.write(
         dataframe=data.referrers_snapshot_df,
         sort_by=["dt", "referrer"],
     )
@@ -33,24 +33,24 @@ def execute_pull_traffic():
 
 
 def execute_pull_activity():
-    data: GithubActivityData = GithubActivityData.fetch(closed_items_last_n_days=10)
+    data: GithubActivityData = GithubActivityData.fetch()
 
-    GithubDataset.PRS.write(
+    Github.PRS.write(
         dataframe=data.prs,
         sort_by=["repo", "number", "state", "updated_at"],
     )
 
-    GithubDataset.ISSUES.write(
+    Github.ISSUES.write(
         dataframe=data.issues,
         sort_by=["repo", "number", "state", "updated_at"],
     )
 
-    GithubDataset.PR_COMMENTS.write(
+    Github.PR_COMMENTS.write(
         dataframe=data.pr_comments,
         sort_by=["repo", "pr_number", "updated_at"],
     )
 
-    GithubDataset.PR_REVIEWS.write(
+    Github.PR_REVIEWS.write(
         dataframe=data.pr_reviews,
         sort_by=["repo", "pr_number", "updated_at"],
     )
