+++ tests/op_analytics/cli/pulls/defillama/test_defillama_protocols.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 
 import polars as pl
 
-from op_analytics.cli.subcommands.pulls.defillama import protocols
+from op_analytics.datasources.defillama import protocols
 from op_analytics.coreutils.testutils.inputdata import InputTestData
 
 TESTDATA = InputTestData.at(__file__)
@@ -302,8 +302,8 @@ def test_construct_urls():
     assert slugs == expected_slugs
 
 
-@patch("op_analytics.cli.subcommands.pulls.defillama.protocols.get_data")
-@patch("op_analytics.cli.subcommands.pulls.defillama.protocols.now_date")
+@patch("op_analytics.datasources.defillama.protocols.get_data")
+@patch("op_analytics.datasources.defillama.protocols.now_date")
 @patch("op_analytics.coreutils.partitioned.dailydata.PartitionedWriteManager.write")
 def test_pull_single_protocol_tvl(
     mock_write,
@@ -378,8 +378,8 @@ def test_pull_single_protocol_tvl(
     ]
 
 
-@patch("op_analytics.cli.subcommands.pulls.defillama.protocols.get_data")
-@patch("op_analytics.cli.subcommands.pulls.defillama.protocols.now_date")
+@patch("op_analytics.datasources.defillama.protocols.get_data")
+@patch("op_analytics.datasources.defillama.protocols.now_date")
 @patch("op_analytics.coreutils.partitioned.dailydata.PartitionedWriteManager.write")
 def test_pull_all_protocol_tvl(
     mock_write,
