+++ tests/op_analytics/cli/pulls/defillama/test_defillama_stablecoins.py
@@ -4,7 +4,7 @@ import pytest
 from op_analytics.coreutils.testutils.inputdata import InputTestData
 
 
-from op_analytics.cli.subcommands.pulls.defillama import stablecoins
+from op_analytics.datasources.defillama import stablecoins
 
 TESTDATA = InputTestData.at(__file__)
 
@@ -109,7 +109,7 @@ def test_process_breakdown_stables():
     assert excinfo.value.args == ("pegType",)
 
 
-@patch("op_analytics.cli.subcommands.pulls.defillama.stablecoins.get_data")
+@patch("op_analytics.datasources.defillama.stablecoins.get_data")
 @patch("op_analytics.coreutils.partitioned.dailydata.PartitionedWriteManager.write")
 def test_pull_stables_single_stablecoin(
     mock_write,
@@ -196,7 +196,7 @@ def test_pull_stables_single_stablecoin(
     ]
 
 
-@patch("op_analytics.cli.subcommands.pulls.defillama.stablecoins.get_data")
+@patch("op_analytics.datasources.defillama.stablecoins.get_data")
 @patch("op_analytics.coreutils.partitioned.dailydata.PartitionedWriteManager.write")
 def test_pull_stables_multiple_stablecoins(
     mock_write,
@@ -300,9 +300,7 @@ def test_pull_stables_multiple_stablecoins(
 
 def test_pull_stables_no_valid_ids():
     # Test pull_stables with invalid stablecoin_ids (should raise ValueError)
-    with patch(
-        "op_analytics.cli.subcommands.pulls.defillama.stablecoins.get_data"
-    ) as mock_get_data:
+    with patch("op_analytics.datasources.defillama.stablecoins.get_data") as mock_get_data:
         mock_get_data.return_value = sample_summary
         with pytest.raises(ValueError) as excinfo:
             stablecoins.pull_stablecoins(symbols=["NOEXIST"])
@@ -311,9 +309,7 @@ def test_pull_stables_no_valid_ids():
 
 def test_pull_stables_missing_pegged_assets():
     # Test pull_stables when 'peggedAssets' is missing (should raise KeyError)
-    with patch(
-        "op_analytics.cli.subcommands.pulls.defillama.stablecoins.get_data"
-    ) as mock_get_data:
+    with patch("op_analytics.datasources.defillama.stablecoins.get_data") as mock_get_data:
         mock_get_data.return_value = {}
         with pytest.raises(KeyError) as excinfo:
             stablecoins.pull_stablecoins()
@@ -349,9 +345,7 @@ def test_process_breakdown_stables_optional_metadata():
 
 def test_pull_stables_empty_metadata_df():
     # Test pull_stables when metadata_df is empty (should raise ValueError)
-    with patch(
-        "op_analytics.cli.subcommands.pulls.defillama.stablecoins.get_data"
-    ) as mock_get_data:
+    with patch("op_analytics.datasources.defillama.stablecoins.get_data") as mock_get_data:
         mock_get_data.side_effect = [
             sample_summary,  # Summary data
             # Data that causes process_breakdown_stables to return empty DataFrames
@@ -373,9 +367,7 @@ def test_pull_stables_empty_metadata_df():
 
 def test_pull_stables_empty_breakdown_df():
     # Test pull_stables when breakdown_df is empty (should raise ValueError)
-    with patch(
-        "op_analytics.cli.subcommands.pulls.defillama.stablecoins.get_data"
-    ) as mock_get_data:
+    with patch("op_analytics.datasources.defillama.stablecoins.get_data") as mock_get_data:
         mock_get_data.side_effect = [
             sample_summary,  # Summary data
             {
