+++ tests/op_analytics/cli/pulls/defillama/test_stablecoin_balance.py
@@ -1,6 +1,6 @@
 from decimal import Decimal
 
-from op_analytics.cli.subcommands.pulls.defillama.stablecoins import safe_decimal
+from op_analytics.datasources.defillama.stablecoins import safe_decimal
 
 
 def test():
